Fix reservoirs nudging animals out of the way
drawing reservoir ghosts uses a different code path than drawing ghosts for most other buildings, because of the draggable behavior
Add code to allow reservoirs to nudge animals out of the way of the build radius, like with other buildings.
Notice that when drawing a reservoir ghost, the 'tile' refers to the center tile of the reservoir, while for most other buildings it is the top-most (according to current view orientation) tile. We have to correct that when checking for obstructions.
Actually the reservoir code checks for obstructions twice, once in draw_draggable_reservoir which does a lot of other calculations, and (now) once in draw_single_reservoir() only for the purpose of animal nudging. I don't think it's going to kill performance, but it's worth noting that we are looping twice and likely can simplify in the future